Experimental fix for blurry textures
This change removes the distance based bias (which is a large contributor to premature blurring) but adds a check using the importance factor to give some balance.
importance should be between 0 and 1 and is higher when the texture is facing the camera, lower when it is side on,
The unimportance setting defines the cutiff vaklue below which we'll consider the textures worth "down scaling" by the bias factor.