Adjust Texture Deletion back to a Grace Period and Test non-Zero Discard loading
After some HEATED DISCUSSIONS, I have found the error in my ways and decided that a grace period on texture deletions are a good idea. Textures will be set inactive for 60 seconds if not found on a face and deleted 60 seconds after that.

Also, testing some code to load in textures without decoding all the way to 0 discard every time. Hopefully will increase scene load times and reduce memory usage.