Avatar Loading Checks Made Independ from Texture Updates
Refactored the avatar mesh loading checks to be independent from the texture update process, where it was previously in LLViewerTextureList::updateImageDecodePriority.

Also made the wait time for the attachment check to be dependant on the number of texture fetches active in relation to the average FPS so that it will adapt to scene complexity and computer's performance.