Texture Processing Fixes
Missed a spot in llViewerTexture for Boost_HIGH and added some fetch qualifications to try to reduce requests during camera movements or high texture load.

Refactored a few things in updateImageDecodePriority, added Particle Boost Level increase and set VSize to a minimum of 64 in face loop because some faces do not seem to have texture entries early on in processing (noticed this with the moon texture).

Re-added Timer limiting to texture creation for now but may remove again later.

Changed batch size to be by time remaining and extended qualifications for secondary fetch loop.