Cleanup Texture Animation Loop
Inside of LLVolumeGeometryManager::rebuildGeom(LLSpatialGroup* group) is a loop that goes through the faces of an object that had another loop inside it also going through a set of faces in the object.

So, moved the vobj->mTextureAnimp loop outside the drawablep->face loop to make it O(n) again.

Also, did a minor change of the variable names in the mTextureAnimp loop so they don't match previous face variables (facep -> facea).