Reduce AGENT_UPDATES_PER_SECOND to 20
This value was originally only 10 several months ago before it was changed in Firestorm and Alchemy to 125.

I think this might be causing the networking issues recently with attachments and such, so reducing this back down to 20.

The 125 number is pretty unreasonable, anyway. If the average user can only turn at most 60 degrees in a second, 125 updates would be 0.48 degrees per update versus 3 degrees an update at 20 (or 6 at 10).